refactor(types): context-menu
